fixed all remaining compiler warnings added -Werror for Debug and Debug_Binonly build mode
git-svn-id: https://bibletime.svn.sf.net/svnroot/bibletime/trunk@718 ba6c7691-6512-0410-bf18-b6da21b6447f

+++ bibletime/src/frontend/searchdialog/analysis/csearchanalysisscene.cpp
@@ -94,7 +94,7 @@ void CSearchAnalysisScene::analyse(ListCSwordModuleInfo modules) {
 		LEGEND_WIDTH, LEGEND_INNER_BORDER*2 + ITEM_TEXT_SIZE*numberOfModules + LEGEND_DELTAY*(numberOfModules-1) );
 	m_legend->show();
 
-	int xPos = LEFT_BORDER + m_legend->rect().width() + SPACE_BETWEEN_PARTS;
+	int xPos = (int)(LEFT_BORDER + m_legend->rect().width() + SPACE_BETWEEN_PARTS);
 	int moduleIndex = 0;
 	m_maxCount = 0;
 	int count = 0;
